+++ test/Fixtures/Mysql/SimpleTestCases/QuotedTableName/database.sql
@@ -1,0 +1,5 @@
+CREATE TABLE `user profile` (
+  `id` int(10) NOT NULL,
+  `user_id` int(10) NOT NULL,
+  PRIMARY KEY (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;